03_COMMODITY_MODULES/cement/atlas/requirements.txt
# Core data processing
duckdb>=1.1.0
pandas>=2.1.0
pyarrow>=14.0.0

# CLI
click>=8.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0

# Entity resolution
rapidfuzz>=3.5.0

# Excel handling (for archival work)
openpyxl>=3.1.0

# Future additions (commented for now)
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
# anthropic>=0.28.0
